+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 
 ## [Unreleased]
-- In IDA, always attempt to `set_bytes`
+- In IDA, check if the address is initialized or is in a valid segment to `set_bytes`
 - For Ghidra, if an address is uninitialized when `set_bytes` is called, attempt to initialize the address if `settings.ghidra.initialize_on_write` is `True`